refactor!: add Kinematics info structure
Undo previous changes to the kinematics interface and use a new data
structure instead, ParticleReactionKinematicsInfo.
Compose the HelicityKinematics of the new data structure.
With pythons duck typing the interface of generate can still accept
Kinematics instances, even though the data structure is not part of the
Kinematics interface. A runtime error will occur if the Kinematics
instance does not have the data structure.